Start refactoring to get rid of WorkingOnResolvers
Right now the typechecker needs a mutable reference to the list of modules, as it was changing the typ values within. With the change to HM we don't actually need this, as we can apply all newly formed types in one fell swoop. Factoring out the bit that actually applies the new types is the start of this. Do more tomorrow